+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/impl/CertificationCriterionDAOImpl.java
@@ -13,12 +13,14 @@ import java.util.List;
 
 import javax.persistence.Query;
 
+import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Repository;
 
 @Repository("certificationCriterionDAO")
 public class CertificationCriterionDAOImpl extends BaseDAOImpl implements CertificationCriterionDAO {
 	
 	@Override
+	@CacheEvict(value="searchOptionsCache", allEntries=true)
 	public CertificationCriterionDTO create(CertificationCriterionDTO dto) throws EntityCreationException, EntityRetrievalException {
 		
 		CertificationCriterionEntity entity = null;
@@ -55,6 +57,7 @@ public class CertificationCriterionDAOImpl extends BaseDAOImpl implements Certif
 	}
 
 	@Override
+	@CacheEvict(value="searchOptionsCache", allEntries=true)
 	public CertificationCriterionDTO update(CertificationCriterionDTO dto) throws EntityRetrievalException, EntityCreationException {
 		
 		CertificationCriterionEntity entity = this.getEntityById(dto.getId());;
@@ -79,6 +82,7 @@ public class CertificationCriterionDAOImpl extends BaseDAOImpl implements Certif
 	}
 	
 	@Override
+	@CacheEvict(value="searchOptionsCache", allEntries=true)
 	public void delete(Long criterionId) {
 		
 		Query query = entityManager.createQuery("UPDATE CertificationCriterionEntity SET deleted = true WHERE certification_criterion_id = :entityid");
@@ -121,6 +125,7 @@ public class CertificationCriterionDAOImpl extends BaseDAOImpl implements Certif
 		return new CertificationCriterionDTO(entity);
 	}
 	
+	@CacheEvict(value="searchOptionsCache", allEntries=true)
 	private void create(CertificationCriterionEntity entity) {
 		
 		entityManager.persist(entity);
@@ -128,6 +133,7 @@ public class CertificationCriterionDAOImpl extends BaseDAOImpl implements Certif
 		
 	}
 	
+	@CacheEvict(value="searchOptionsCache", allEntries=true)
 	private void update(CertificationCriterionEntity entity) {
 		
 		entityManager.merge(entity);