+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/impl/DeveloperDAOImpl.java
@@ -9,6 +9,7 @@ import javax.persistence.Query;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
@@ -36,6 +37,7 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 	@Autowired ContactDAO contactDao;
 
 	@Override
+	@CacheEvict(value="searchOptionsCache", allEntries=true)
 	@Transactional
 	public DeveloperDTO create(DeveloperDTO dto) throws EntityCreationException, EntityRetrievalException {
 
@@ -118,6 +120,7 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 	}
 
 	@Override
+	@CacheEvict(value="searchOptionsCache", allEntries=true)
 	@Transactional
 	public DeveloperEntity update(DeveloperDTO dto) throws EntityRetrievalException {
 		DeveloperEntity entity = this.getEntityById(dto.getId());
@@ -332,7 +335,8 @@ public class DeveloperDAOImpl extends BaseDAOImpl implements DeveloperDAO {
 		entityManager.persist(entity);
 		entityManager.flush();
 	}
-
+	
+	@CacheEvict(value="searchOptionsCache", allEntries=true)
 	private void update(DeveloperEntity entity) {
 		entityManager.merge(entity);
 		entityManager.flush();