+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/CHPLTestConfig.java
@@ -17,6 +17,7 @@ import org.springframework.context.EnvironmentAware;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.EnableAspectJAutoProxy;
 import org.springframework.context.annotation.FilterType;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.core.env.Environment;
@@ -44,6 +45,8 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 import com.github.springtestdbunit.bean.DatabaseConfigBean;
 import com.github.springtestdbunit.bean.DatabaseDataSourceConnectionFactoryBean;
 
+import gov.healthit.chpl.util.MethodLogger;
+
 
 @Configuration
 //@EnableWebSecurity
@@ -51,6 +54,7 @@ import com.github.springtestdbunit.bean.DatabaseDataSourceConnectionFactoryBean;
 @PropertySource("classpath:/environment.test.properties")
 @EnableCaching
 @EnableTransactionManagement
+@EnableAspectJAutoProxy
 @ComponentScan(basePackages = {"gov.healthit.chpl.**"}, excludeFilters = {@ComponentScan.Filter(type = FilterType.ANNOTATION, value = Configuration.class)})
 public class CHPLTestConfig implements EnvironmentAware {
 	
@@ -171,8 +175,6 @@ public class CHPLTestConfig implements EnvironmentAware {
 	public EhCacheFactoryBean ehCacheFactoryBean(){
 		EhCacheFactoryBean bean = new EhCacheFactoryBean();
 		bean.setCacheManager(ehCacheCacheManager().getObject());
-		//bean.setCacheName("aclCache");
-		//bean.setCacheName(env.getProperty("authAclCacheName"));
 		return bean;
 	}