+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/web/controller/DeveloperControllerTest.java
@@ -2,6 +2,9 @@ package gov.healthit.chpl.web.controller;
 
 import static org.junit.Assert.assertTrue;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -24,6 +27,8 @@ import gov.healthit.chpl.auth.permission.GrantedPermission;
 import gov.healthit.chpl.auth.user.JWTAuthenticatedUser;
 import gov.healthit.chpl.dao.EntityCreationException;
 import gov.healthit.chpl.dao.EntityRetrievalException;
+import gov.healthit.chpl.domain.Developer;
+import gov.healthit.chpl.domain.UpdateDevelopersRequest;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = { gov.healthit.chpl.CHPLTestConfig.class })
@@ -35,6 +40,9 @@ import gov.healthit.chpl.dao.EntityRetrievalException;
 public class DeveloperControllerTest {
 	@Autowired
 	DeveloperController developerController = new DeveloperController();
+	
+	@Autowired
+	SearchViewController searchViewController = new SearchViewController();
 
 	private static JWTAuthenticatedUser adminUser;
 	
@@ -69,6 +77,53 @@ public class DeveloperControllerTest {
 				+ " millis or " + getDevelopersElapsedSeconds + " seconds");
 		assertTrue("DeveloperController.getDevelopers() should complete within 3 seconds but took " + getDevelopersTimeLength
 				+ " millis or " + getDevelopersElapsedSeconds + " seconds", getDevelopersElapsedSeconds < 3);
+	}
+	
+	/** Description: Tests that the updateDevelopers API call triggers the searchOptionsCache to evict/refresh
+	 * 
+	 * Expected Result: Executing the updateDevelopers method should cause the searchOptions cache to be refreshed
+	 */
+	@Transactional
+	@Rollback(true)
+	@Test
+	public void test_updateDevelopers_triggersSearchOptionsCacheEvict() throws EntityRetrievalException, JsonProcessingException, EntityCreationException{
+		SecurityContextHolder.getContext().setAuthentication(adminUser);
+		long startTime = System.currentTimeMillis();
+		Boolean required = true;
+		// Cache SearchOptions
+		searchViewController.getPopulateSearchData(required);
+		long endTime = System.currentTimeMillis();
+		long timeLength = endTime - startTime;
+		double elapsedSeconds = timeLength / 1000.0;
+		assertTrue("search options should take longer than 100 ms to call the first time but took " + elapsedSeconds + 
+				" and " + timeLength + "ms", timeLength > 100L);
+		
+		// Verify that SearchOptions is cached
+		startTime = System.currentTimeMillis();
+		searchViewController.getPopulateSearchData(required);
+		endTime = System.currentTimeMillis();
+		timeLength = endTime - startTime;
+		elapsedSeconds = timeLength / 1000.0;
+		assertTrue("search options should now be cached and take < 100 ms but took " + timeLength + "ms", timeLength < 100L);
+		
+		Developer developer = new Developer();
+		developer.setDeveloperId(-1L);
+		
+		UpdateDevelopersRequest updateDevelopersRequest = new UpdateDevelopersRequest();
+		updateDevelopersRequest.setDeveloper(developer);
+		List<Long> developerIds = new ArrayList<Long>();
+		developerIds.add(-1L);
+		updateDevelopersRequest.setDeveloperIds(developerIds);
+		
+		// Evict SearchOptions cache and verify that cache is cleared
+		developerController.updateDeveloper(updateDevelopersRequest);
+		
+		startTime = System.currentTimeMillis();
+		searchViewController.getPopulateSearchData(required);
+		endTime = System.currentTimeMillis();
+		timeLength = endTime - startTime;
+		elapsedSeconds = timeLength / 1000.0;
+		assertTrue("search options should no longer be cached but took " + timeLength + "ms", timeLength > 100L);
 	}
 	
 }